Draws samples from the posterior of the factorial model for categorical data.

Draws samples from the posterior of the factorial model for categorical data.

Usage

fitfatcat(
  y,
  q,
  nit,
  burnin = 0,
  lag = 1,
  init_beta = 0.5,
  init_sigma2 = 2,
  C0 = 1e+05,
  a = 0.001,
  b = 0.001,
  sdpropbeta = 0.05,
  sdpropbeta2 = 0.075,
  sdpropf = 0.4,
  sdpropsigma2 = 0.075,
  dist = c("probit"),
  quiet = F,
  alpha = NULL
)

Arguments

y

A matrix of integers with dimension j x n, where j is the number of variables and n the number of observations.

q

An integer with the number of factors.

nit

An integer with the number of iterations.

burnin

An integer with the number of initial iterations to be ignored.

lag

An integer with the lag of the iterations taken. ``lag = 1`` means take every iteration. ``lag = n`` means take one iteration every n iterations.

init_beta

An integer with the initial value of the first column of beta.

init_sigma2

An integer with the initial value of sigma2. All elements of the sigma2 vector will be initialized with the same value.

C0

An integer with the variance of the prior for beta.

a, b

An integer with the shape and scale parameters of the prior of sigma2.

sdpropbeta

An integer with the standard deviation of the proposal of betas that are unrestricted.

sdpropbeta2

An integer with the standard deviation of the proposal of betas that are restricted to be greater than or equal to zero.

sdpropf

An integer with the standard deviation of the proposal of factors.

sdpropsigma2

An integer with the standard deviation of the proposal of sigma2.

dist

An string with the name of the error distribution. Currently, only `probit` is accepted.

quiet

If TRUE, prevents the display of progress bar.

alpha

An vector with the true value of alpha. Usually, starts with -Inf and ends with +Inf.

Value

A posterior object with a sample from the posterior distribution.

Examples

post = fitfatcat(ordered_data$y, q = 2, nit = 10000,lag = 9, burnin = 1000, quiet = TRUE)
beta_post = posterior::subset_draws(post, variable = "beta", regex = TRUE)
posterior::summarise_draws(beta_post)
#> # A tibble: 10 × 10
#>    variable   mean median     sd    mad      q5   q95  rhat ess_bulk ess_tail
#>    <chr>     <num>  <num>  <num>  <num>   <num> <num> <num>    <num>    <num>
#>  1 beta_11  1.04   1.03   0.0718 0.0796  0.924  1.16   1.02     49.3    187. 
#>  2 beta_21  0.0189 0.0166 0.0709 0.0629 -0.0942 0.146  1.01     57.8     94.9
#>  3 beta_31  0.988  0.985  0.0710 0.0731  0.882  1.11   1.02     78.7    260. 
#>  4 beta_41  0.0911 0.0897 0.0700 0.0615 -0.0190 0.219  1.00     65.4     96.7
#>  5 beta_51  0.636  0.634  0.0730 0.0724  0.522  0.761  1.01     99.9    199. 
#>  6 beta_12  0      0      0      0       0      0     NA        NA       NA  
#>  7 beta_22  0.948  0.945  0.0675 0.0674  0.846  1.06   1.01    114.     277. 
#>  8 beta_32  0.101  0.101  0.0374 0.0365  0.0403 0.162  1.00    297.     538. 
#>  9 beta_42  0.959  0.961  0.0652 0.0706  0.850  1.07   1.04     61.6    176. 
#> 10 beta_52  0.646  0.646  0.0565 0.0548  0.551  0.742  1.00    223.     480.